main_server/requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==3.2.18
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.5.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
whitenoise==6.2.0
psycopg2-binary
coverage